Permit host filesystem access for Thonny
This is on par with all of the other IDE and code editor Flatpaks.
The sandbox is not the best experience for development and has some limitations.
The way the sandbox exposes files is secure, but not ideal for code editors where the path to the file is obscure.
It's much easier for things to be mounted where they are expected.
Additionally, mounted media needs to be accessible to Thonny, too.