Make it possible to install and use Python packages from Thonny
The installation of user packages is isolated inside the Flatpak. This prevents contamination of user installed packages on the host. Filesystem access is now only permitted for the mount directories and the home directory. This allows for accessing mounted media by Thonny, which is necessary, while permitting less access from within the sandbox.

Ideally the Flatpak would not allow access to the home directory directly. Portals should be used, but Thonny doesn't permitting opening entire directories. This means using multiple Python files is extremely difficult when the sandbox is fully in effect. Work upstream in Thonny could make this possible.